% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rc_coerceData.R
\name{coerceData}
\alias{coerceData}
\title{coerceData}
\usage{
coerceData(df)
}
\arguments{
\item{df}{Data object (any object that can be coerced to a data frame)}
}
\value{
A data frames, which is DO as data.frames
}
\description{
coerceData
}
