% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pd_coerceFactorsToChar.R
\name{coerceFactorsToChar}
\alias{coerceFactorsToChar}
\title{coerceFactorsToChar: convert all factor type fields to characters}
\usage{
coerceFactorsToChar(df)
}
\arguments{
\item{df}{Input dataframe}
}
\value{
\code{df} with factor fields converted to character type
}
\description{
coerceFactorsToChar: convert all factor type fields to characters
}
