% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cd_collapseClasses.R
\name{collapseClasses}
\alias{collapseClasses}
\title{collapseClasses. Collapse the classes of an object to a single string}
\usage{
collapseClasses(x)
}
\arguments{
\item{x}{any object}
}
\value{
a string listing the classes of x, separated by commas
}
\description{
collapseClasses. Collapse the classes of an object to a single string
}
