% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rco_createCleaningInfo.R
\name{createCleaningInfo}
\alias{createCleaningInfo}
\title{Converts cleaning info into a format consumable by updateReviewObject.}
\usage{
createCleaningInfo(revObj, cleaningInfo)
}
\arguments{
\item{revObj}{dataReviewRobject to be updated}

\item{cleaningInfo}{list of cleaning information}
}
\value{
\code{revObj} updated dataReviewRobject
}
\description{
Converts cleaning info into a format consumable by updateReviewObject.
}
