% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rco_createFreq.R
\name{createFreq}
\alias{createFreq}
\title{Takes the raw info for the freq block of the output
and puts it in a format usable by the updateReviewObject
function}
\usage{
createFreq(revObj, df)
}
\arguments{
\item{revObj}{dataReviewRobject to be updated}

\item{df}{cleaned data set returned as list element 'coerceData' from function prepareData}
}
\value{
\code{dataReviewRobject}
}
\description{
Takes the raw info for the freq block of the output
and puts it in a format usable by the updateReviewObject
function
}
