% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rco_createFreq.R
\name{createFreqList}
\alias{createFreqList}
\title{Create freq list: a list of data frames. One list item for each original dataset column.
The data frame in each list item includes the columns contents sorted by frequency}
\usage{
createFreqList(df)
}
\arguments{
\item{df}{dataframe, output from prepareData}
}
\value{
An dataReviewR freq list
}
\description{
Create freq list: a list of data frames. One list item for each original dataset column.
The data frame in each list item includes the columns contents sorted by frequency
}
