% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rco_createMeta.R
\name{createMeta}
\alias{createMeta}
\title{Takes the raw info for the meta block of the output
and puts it in a format usable by the updateReviewObject
function}
\usage{
createMeta(dataReviewRobject, df, arguments, timestamp, roundDigits)
}
\arguments{
\item{dataReviewRobject}{Object of class dataReviewRobject}

\item{df}{Data set passed in to the dataReviewR function}

\item{arguments}{Collection of arguments passed to Review object with labels that match the dataReviewR arg definitions}

\item{timestamp}{Timestamp}

\item{roundDigits}{The number of digits to round to, using \link[base]{round}}
}
\value{
\code{dataReviewRobject}
}
\description{
Takes the raw info for the meta block of the output
and puts it in a format usable by the updateReviewObject
function
}
