% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pf_createReviewObject.R
\name{createReviewObject}
\alias{createReviewObject}
\title{Generates an empty list of the correct class to store results}
\usage{
createReviewObject()
}
\value{
A list of class dataReviewRObject
}
\description{
Generates an empty list of the correct class to store results
}
