% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rco_createSkimInfo.R
\name{createSkimInfo}
\alias{createSkimInfo}
\title{Converts skim info into a format consumable by updateReviewObject.}
\usage{
createSkimInfo(revObj, cleaned_df)
}
\arguments{
\item{revObj}{dataReviewRobject to be updated}

\item{cleaned_df}{cleaned dataframe returned by prepareDate (the coerceData portion)}
}
\value{
\code{revObj} updated dataReviewRobject
}
\description{
Converts skim info into a format consumable by updateReviewObject.
}
