% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pd_executeCoercions.R
\name{executeCoercions}
\alias{executeCoercions}
\title{executeCoercions:}
\usage{
executeCoercions(df, whitespace_trim = TRUE, clean_var_names = TRUE)
}
\arguments{
\item{df}{Input dataframe}

\item{whitespace_trim}{User defined boolean for whether leading/trailing white space is trimmed in strings (TRUE / FALSE)}

\item{clean_var_names}{Boolean.}
}
\value{
\code{out} list containing 3 data frames df, data_types, data_names

\code{df} Dataframe with factor fields converted to character type and white space trimming (if option is selected by the user)

\code{data_types} Dataframe with field types before and after cleaning for both df and DFB
}
\description{
executeCoercions:
}
