% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rco_createSkimInfo.R
\name{get_df_key_names}
\alias{get_df_key_names}
\title{Returns character vector of the dataset columns that appear to be keys (i.e. the column contains unique values)}
\usage{
get_df_key_names(df_skim, df_nrow)
}
\arguments{
\item{df_skim}{skim output from createSkimInfo}

\item{df_nrow}{integer. Number of rows in the data set}
}
\value{
\code{key_var_names} Character vector of dataset key columns (i.e. column names where column data is unique)
}
\description{
Returns character vector of the dataset columns that appear to be keys (i.e. the column contains unique values)
}
