% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rco_createCleaningInfo.R
\name{getNames}
\alias{getNames}
\title{variables new and old names}
\usage{
getNames(namesDF)
}
\arguments{
\item{namesDF}{Dataframe of type information from the executeCoercion function}
}
\value{
\code{coercedT}
}
\description{
variables new and old names
}
