% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rc_isRReviewObject.R
\name{is.dataReviewRobject}
\alias{is.dataReviewRobject}
\title{Check object is of class dataReviewRobject}
\usage{
is.dataReviewRobject(x)
}
\arguments{
\item{x}{An object}
}
\value{
A boolean: TRUE if object is class dataReviewRobject and FALSE if not
}
\description{
Check object is of class dataReviewRobject
}
