% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rco_updateReviewObject_Meta.R
\name{metaDataInfo}
\alias{metaDataInfo}
\title{Creates a list of info about the dataframe.}
\usage{
metaDataInfo(name, df)
}
\arguments{
\item{name}{The variable name of the df from the dataReviewR function call}

\item{df}{A data frame}
}
\value{
\code{dfInfo} A list of info about the data frame
}
\description{
Creates a list of info about the dataframe.
}
