% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pf_processFlow.R
\name{processFlow}
\alias{processFlow}
\title{processFlow Handles the process flow for the whole package}
\usage{
processFlow(df, roundDigits, trimChars, clean_var_names, argsIn)
}
\arguments{
\item{df}{Dataframe. The data frame to be reviewed}

\item{roundDigits}{Integer. If NA, numerics are not rounded before comparison. If /code{roundDigits} is specified, numerics are
rounded to /code{roundDigits} decimal places using \link[base]{round}.}

\item{trimChars}{Boolean. Do we trim characters before comparing?}

\item{clean_var_names}{Boolean.}

\item{argsIn}{The arguments that were passed to the main data_reviewer R function

(without producing a dataCompareR object). Designed to improve performance for large datasets.}
}
\value{
\code{dataCompareRObject} containing details of the comparison
}
\description{
processFlow Handles the process flow for the whole package
}
