% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/out_saveReport.R
\name{saveReport}
\alias{saveReport}
\title{Save a report based on a dataReviewR object}
\usage{
saveReport(
  reviewObject,
  reportName,
  reportLocation = ".",
  HTMLReport = TRUE,
  showInViewer = TRUE,
  stylesheet = NA,
  printAll = FALSE,
  head_tail_n = 20,
  max_table_length = 50,
  ...
)
}
\arguments{
\item{reviewObject}{a dataReviewR object.}

\item{reportName}{String. The name of the report. Reports will be saved as reportName.Rmd and (optionally) reportName.html in
\code{reportLocation}}

\item{reportLocation}{String. Location to save reports specified by the user. The R markdown and (optionally) HTML reports will
be saved in this area}

\item{HTMLReport}{Boolean. Option to output html report.}

\item{showInViewer}{Boolean. Does the html report open automatically in the viewer?}

\item{stylesheet}{String. Optional link to customised css stylesheet}

\item{printAll}{Boolean. If TRUE, all mis-matches in the data are printed to the file. This acts as a shortcut}

\item{head_tail_n}{Integer. The number of rows to keep in the head and tail in the output sample data}

\item{max_table_length}{Integer. The maximum number of most frequent entries to show in output summaries
to get all mismatches in the report, reviewd to passing the number in \code{mismatchCount}. When TRUE, overrides the
\code{mismatchCount} field passed via ellipses}

\item{...}{Optional arguments which will be passed to \code{summary}, for example \code{mismatchCount}}
}
\description{
Saves R markdown and HTML reports in the area specified by the user.

Uses knitr and markdown to create reports. Reports have the extensions .Rmd or .html.
By default the \code{table.css} style sheet is used for format the html output.
}
\seealso{
Other dataQual.functions: 
\code{\link{rReview}()},
\code{\link{summary.dataReviewRobject}()}
}
\concept{dataQual.functions}
