% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pd_trimCharVars.R
\name{trimCharVars}
\alias{trimCharVars}
\title{trimCharVars: trim white spaces in character variables from an input dataframe}
\usage{
trimCharVars(df)
}
\arguments{
\item{df}{Input dataframe}
}
\value{
\code{df} with preceding and trailing white spaces removed from character fields
}
\description{
trimCharVars: trim white spaces in character variables from an input dataframe
}
