% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rco_updateCompareObject_CleaningInfo.R
\name{updateReviewObject.cleaninginfo}
\alias{updateReviewObject.cleaninginfo}
\title{Updates cleaning info in the revare object}
\usage{
\method{updateReviewObject}{cleaninginfo}(x, revObj)
}
\arguments{
\item{x}{list of type cleaninginfo with data types}

\item{revObj}{dataReviewRobject to be updated}
}
\value{
\code{revObj} updated dataReviewRobject
}
\description{
Updates cleaning info in the revare object
}
