% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rco_updateReviewObject_Freq.R
\name{updateReviewObject.freq}
\alias{updateReviewObject.freq}
\title{Adds a freq block to the output}
\usage{
\method{updateReviewObject}{freq}(x, revObj)
}
\arguments{
\item{x}{List of class 'freq' with column matching info}

\item{revObj}{dataReviewRobject instance to be updated}
}
\value{
\code{revObj} Updated dataReviewRobject
}
\description{
Adds a freq block to the output
}
