% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rco_updateReviewObject_Meta.R
\name{updateReviewObject.meta}
\alias{updateReviewObject.meta}
\title{Takes raw info for meta and adds it to the review object}
\usage{
\method{updateReviewObject}{meta}(x, compObj)
}
\arguments{
\item{x}{List of class 'meta' with data related to meta}

\item{compObj}{dataReviewRobject to be appended}
}
\value{
\code{compObj} dataReviewRobject updated with meta block
}
\description{
Takes raw info for meta and adds it to the review object
}
