% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rco_updateReviewObject.R
\name{updateReviewObject}
\alias{updateReviewObject}
\title{Generic function for updating a review object with
information passed to it, that has methods based on the class
of the info argument.}
\usage{
updateReviewObject(x, revObj)
}
\arguments{
\item{x}{Object of information with classes related to the relevant section of the dataReviewRobject}

\item{revObj}{dataReviewRobject to be updated}
}
\value{
revObj Updated dataReviewRobject
}
\description{
Generic function for updating a review object with
information passed to it, that has methods based on the class
of the info argument.
}
