% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rco_updateReviewObject_Skim.R
\name{updateReviewObject.skim}
\alias{updateReviewObject.skim}
\title{Takes raw info for skim and adds it to the review object}
\usage{
\method{updateReviewObject}{skim}(x, revObj)
}
\arguments{
\item{x}{List of class 'skim' with data related to skim}

\item{revObj}{dataReviewRobject to be appended}
}
\value{
\code{revObj} dataReviewRobject updated with skim block
}
\description{
Takes raw info for skim and adds it to the review object
}
