% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rc_validateArguments.R
\name{validateArguments}
\alias{validateArguments}
\title{validateArguments}
\usage{
validateArguments(roundDigits = NA, coerceCols = TRUE, trimChars = TRUE)
}
\arguments{
\item{roundDigits}{Number of decimal places to round numeric data column}

\item{coerceCols}{Boolean - do we coerce columns names?}

\item{trimChars}{Boolean. If true, strings and factors have whitespace trimmed before comparison.}
}
\value{
Nothing. Errors if any parameters are invalid.
}
\description{
validateArguments
}
