% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vd_validateData.R
\name{validateData}
\alias{validateData}
\title{validateData : routine to validate the input data}
\usage{
validateData(df)
}
\arguments{
\item{df}{a data frame}
}
\value{
None. Stops if error.
}
\description{
validateData : routine to validate the input data
}
