% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rc_rReview.R
\name{warnLargeData}
\alias{warnLargeData}
\title{Warn users if the calculation is likely to be slow}
\usage{
warnLargeData(nrow_df, ncol_df)
}
\arguments{
\item{nrow_df}{number of rows in data frame}

\item{ncol_df}{number of columns in data frame}
}
\value{
Nothing
}
\description{
Checks if there are more than 20E6 elements for comparison. If there are, spits out a warning
message that the calculation may run slowly
}
